Refactoring to make it easier to reuse
medic/medic#5604
